pyproject.toml
[project]
name = "gh-cloner"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "pygithub>=2.6.1",
    "requests>=2.32.4",
]
